Struct libnotify::Notification

pub struct Notification(_, _);

Notification represents a passive pop-up notification. It can contain summary text, body text, and an icon, as well as hints specifying how the notification should be presented. The notification is rendered by a notification daemon, and may present the notification in any number of ways. As such, there is a clear separation of content and presentation, and this API enforces that.

Methods

impl Notification

fn new<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
summary: &str,
body: P,
icon: Q
) -> Notification

Creates a new Notification. The summary text is required, but all other parameters are optional.

The required summary text.

body

The optional body text.

icon

The optional icon theme icon name or filename.

Returns

The new Notification.

fn close(&self) -> Result<(), Error>

Synchronously tells the notification server to hide the notification on the screen.

Returns

Ok(()) on success, or Err(err) on error

fn show(&self) -> Result<(), Error>

Tells the notification server to display the notification on the screen.

Returns

Ok(()) on success, or Err(err) on error

fn set_hint(&self, key: &str, value: Option<Variant>)

Sets a hint for key with value value. If value is None, a previously set hint for key is unset.

If value is floating, it is consumed.

key

the hint key

value

fn set_image_from_pixbuf(&self, pixbuf: &Pixbuf)

Sets the image in the notification from a gdk_pixbuf::Pixbuf.

pixbuf

The image.

fn clear_hints(&self)

Clears all hints from the notification.

fn set_app_name<'a, P: Into<Option<&'a str>>>(&self, app_name: P)

Sets the application name for the notification. If this function is not called or if app_name is None, the application name will be set from the value used in notify_init or overridden with notify_set_app_name.

app_name

the localised application name

fn set_category(&self, category: &str)

Sets the category of this notification. This can be used by the notification server to filter or display the data in a certain way.

category

The category.

fn set_timeout(&self, timeout: i32)

Sets the timeout of the notification. To set the default time, pass NOTIFY_EXPIRES_DEFAULT as timeout. To set the notification to never expire, pass NOTIFY_EXPIRES_NEVER.

Note that the timeout may be ignored by the server.

timeout

The timeout in milliseconds.

fn set_urgency(&self, urgency: Urgency)

Sets the urgency level of this notification.

See: Urgency

urgency

The urgency level.

fn update<'a, 'b, P: Into<Option<&'a str>>, Q: Into<Option<&'b str>>>(
&self,
summary: &str,
body: P,
icon: Q
) -> Result<(), String>

Updates the notification text and icon. This won't send the update out and display it on the screen. For that, you will need to call Notification::show.

summary

The new required summary text.

body

The optional body text.

icon

The optional icon theme icon name or filename.

Returns

Ok(()) on success, or Err(str) if an invalid parameter was passed